Add ChaiScript headers
Alternatives include possibly using ChaiScript via a git submodule.

This is the quick and dirty approach. The main downside is diverging
from the updates ChaiScript is providing. The chaiscript "master"
branch is guaranteed to be stable, so it could be used as a submodule
base, or you could update manually.

IMPORTANT NOTE: Current ChaiScript "code_cleanups" branch is 2x faster
than current ChaiScript "master". A release will be made soon and
master will be updated.

ALSO: This code is the v5.4.0 tag of ChaiScript, so you know the starting
point.